+++ demis-portal-core-library/src/lib/components/max-height-content-container/max-height-content-container.component.ts
@@ -23,7 +23,7 @@ import { debounceTime, Subject, takeUntil } from 'rxjs';
   selector: 'gem-demis-max-height-content-container',
   template: `
     <div id="gem-demis-max-height-content-container" [style.height]="containerHeight()">
-      <ng-content></ng-content>
+      <ng-content />
     </div>
   `,
   styles: [